Static Libraries and ZERO Warnings

+++ Mango/src/Platform/Windows/WindowsInput.cpp
@@ -27,7 +27,7 @@ namespace Mango {
 		auto window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
-		return { xpos, ypos };
+		return { (float)xpos, (float)ypos };
 	}
 	
 	float WindowsInput::GetMouseXImpl()
